+++ bindings/mnt/pyfiction/src/pyfiction/technology/sidb/simulation/analysis/time_to_solution.cpp
@@ -47,6 +47,8 @@ void time_to_solution(nanobind::module_& m)
     py::class_<fiction::sidb::simulation::analysis::time_to_solution_params>(
         m, "time_to_solution_params", DOC(fiction_sidb_simulation_analysis_time_to_solution_params))
         .def(py::init<>(), "Default constructor.")
+        .def_rw("timeout", &fiction::sidb::simulation::analysis::time_to_solution_params::timeout,
+                DOC(fiction_sidb_simulation_analysis_time_to_solution_params_timeout))
         .def_rw("engine", &fiction::sidb::simulation::analysis::time_to_solution_params::engine,
                 DOC(fiction_sidb_simulation_analysis_time_to_solution_params_engine))
         .def_rw("repetitions", &fiction::sidb::simulation::analysis::time_to_solution_params::repetitions,